remarks were fixed
