+++ src/test/java/Config/ApiConfig.java
@@ -1,3 +1,5 @@
+package Config;
+
 public class ApiConfig {
     public static final String BASE_URL = "https://qa-scooter.praktikum-services.ru/api/v1";
 }
