+++ src/test/java/Config/OrderApi.java
@@ -1,3 +1,7 @@
+package Config;
+
+import Config.ApiConfig;
+import Dto.OrderDto;
 import io.restassured.RestAssured;
 import io.restassured.response.Response;
