--- src/test/java/CourierDto.java
@@ -1,45 +1,0 @@
-
-
-public class CourierDto {
-    private String login;
-    private String password;
-    private String firstName;
-
-    public CourierDto(){
-
-    }
-
-    public CourierDto(String login, String password, String firstName) {
-        this.login = login;
-        this.password = password;
-        this.firstName = firstName;
-    }
-
-    public CourierDto(String login) {
-        this.login = login;
-    }
-
-    public String getLogin() {
-        return login;
-    }
-
-    public void setLogin(String login) {
-        this.login = login;
-    }
-
-    public String getPassword() {
-        return password;
-    }
-
-    public void setPassword(String password) {
-        this.password = password;
-    }
-
-    public String getFirstName() {
-        return firstName;
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
-}
