+++ src/test/java/Dto/CourierDto.java
@@ -1,0 +1,25 @@
+package Dto;
+
+import lombok.Getter;
+import lombok.Setter;
+
+@Getter
+@Setter
+public class CourierDto {
+    private String login;
+    private String password;
+    private String firstName;
+
+    public CourierDto(){
+    }
+
+    public CourierDto(String login, String password, String firstName) {
+        this.login = login;
+        this.password = password;
+        this.firstName = firstName;
+    }
+
+    public CourierDto(String login) {
+        this.login = login;
+    }
+}
