+++ src/test/java/Dto/LoginDto.java
@@ -1,3 +1,5 @@
+package Dto;
+
 public class LoginDto {
     private String login;
     private String password;
